added thread queue

+++ code/lvl5_types.h
@@ -5,12 +5,12 @@ typedef double f64;
 
 typedef char i8;
 typedef short i16;
-typedef int i32;
+typedef long i32;
 typedef long long i64;
 
 typedef unsigned char u8;
 typedef unsigned short u16;
-typedef unsigned int u32;
+typedef unsigned long u32;
 typedef unsigned long long u64;
 
 typedef unsigned char byte;
